+++ crypto/mdc2/mdc2.h
@@ -82,9 +82,9 @@ typedef struct mdc2_ctx_st
 
 
 int MDC2_Init(MDC2_CTX *c);
-int MDC2_Update(MDC2_CTX *c, const unsigned char *data, unsigned long len);
+int MDC2_Update(MDC2_CTX *c, const unsigned char *data, size_t len);
 int MDC2_Final(unsigned char *md, MDC2_CTX *c);
-unsigned char *MDC2(const unsigned char *d, unsigned long n,
+unsigned char *MDC2(const unsigned char *d, size_t n,
 	unsigned char *md);
 
 #ifdef  __cplusplus
